Relocate granulesList.js and update reference in unit test
