+++ packages/api/tests/endpoints/granules/test-bulk-delete.js
@@ -16,7 +16,7 @@ const {
   setAuthorizedOAuthUsers,
 } = require('../../../lib/testUtils');
 const AccessToken = require('../../../models/access-tokens');
-const { granuleIds } = require('./granulesList');
+const { granuleIds } = require('../../helpers/granulesList');
 
 let accessTokenModel;
 let jwtAuthToken;
